requirements.txt
Flask==2.3.2
ultralytics==8.0.196
opencv-python==4.8.0.74
Pillow==10.0.0
numpy==1.24.3

